+++ cypress/integration/middleware/enhanced.spec.ts
@@ -25,4 +25,15 @@ describe('Enhanced middleware', () => {
         .that.includes('This was static but has been transformed in')
     })
   })
+
+  it('adds geo data', () => {
+    cy.request('/api/geo').then((response) => {
+      expect(response.body).to.have.nested.property('headers.x-geo-country')
+      expect(response.body).to.have.nested.property('headers.x-geo-region')
+      expect(response.body).to.have.nested.property('headers.x-geo-city')
+      expect(response.body).to.have.nested.property('headers.x-geo-longitude')
+      expect(response.body).to.have.nested.property('headers.x-geo-latitude')
+      expect(response.body).to.have.nested.property('headers.x-geo-timezone')
+    })
+  })
 })